requirements.txt
Django==3.1.2
freeze==3.0
cx-Oracle==8.0.1
django-widget-tweaks==1.4.8
Pillow==7.2.0
django-extra-views==0.13.0